read version string from LTE_VERSION.

+++ builder/run_builder.sh
@@ -2,10 +2,20 @@
 ABS_SH=`readlink -f $0`
 ABS_DIR=`dirname $ABS_SH`
 
+if [ -z "$LTE_DIR" ]; then
+  echo "LTE_DIR environment not set"
+  exit 1
+fi
+
+if [ -z "$LTE_VERSION" ]; then
+  echo "LTE_VERSION environment not set"
+  exit 1
+fi
+
 sudo -E rm -rf $ABS_DIR/docker_dist
 mkdir $ABS_DIR/docker_dist
 
-sudo -E docker run -v $LTE_DIR:/tmp/lte -v $ABS_DIR/../worker:/tmp/worker -v $ABS_DIR/docker_dist:/tmp/docker_dist lighttransport/lte_builder /tmp/run_internal.sh
+sudo -E docker run -v $LTE_DIR:/tmp/lte -v $ABS_DIR/../worker:/tmp/worker -v $ABS_DIR/docker_dist:/tmp/docker_dist lighttransport/lte_builder /tmp/run_internal.sh $LTE_VERSION
 
 sudo -E cp dockerfile_internal $ABS_DIR/docker_dist/Dockerfile
